HydraPoison-TI/requirements.txt
torch
torchvision
pillow
numpy
regex
unidecode
tqdm